+++ app/_layout.tsx
@@ -1,13 +1,12 @@
 import { Stack } from "expo-router";
-import Toast from 'react-native-toast-message';
+import { NotificationProvider } from "../providers/NotificationProvider";
 
 export default function RootLayout() {
   return (
-    <>
+    <NotificationProvider>
       <Stack screenOptions={{ headerShown: false }}>
         <Stack.Screen name="index" />
       </Stack>
-      <Toast />
-    </>
+    </NotificationProvider>
   );
 }